#108343 add BlockListModel extensions for IPublishedRequest
