+++ GovUk.Frontend.Umbraco/Models/BlockListModelExtensions.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using System.Collections.ObjectModel;
-using System.Linq;
 using Umbraco.Cms.Core;
 using Umbraco.Cms.Core.Models.Blocks;
-using Umbraco.Cms.Core.Routing;
 using Umbraco.Extensions;
 
 namespace GovUk.Frontend.Umbraco.Models
@@ -22,38 +20,6 @@ namespace GovUk.Frontend.Umbraco.Models
         }
 
         /// <summary>
-        /// Recursively find the first block in a request that is bound to a model property using the 'Model property' Umbraco property in the block's settings
-        /// </summary>
-        /// <param name="request"></param>
-        /// <param name="propertyName"></param>
-        /// <returns></returns>
-        /// <exception cref="ArgumentNullException"></exception>
-        public static BlockListItem? FindBlock(this IPublishedRequest request, string propertyName)
-        {
-            if (request?.PublishedContent?.Properties is null)
-            {
-                throw new ArgumentNullException(nameof(request.PublishedContent.Properties));
-            }
-
-            var blockLists = request.PublishedContent.Properties
-                .Where(x => x.PropertyType.EditorAlias == Constants.PropertyEditors.Aliases.BlockList && x.HasValue())
-                .Select(x => x.Value<BlockListModel>(null));
-
-            foreach (var blockList in blockLists)
-            {
-                if (blockList != null)
-                {
-                    var block = blockList.FindBlockByBoundProperty(propertyName);
-                    if (block != null)
-                    {
-                        return block;
-                    }
-                }
-            }
-            return null;
-        }
-
-        /// <summary>
         /// Recursively find the first block in a block list that is bound to a model property using the 'Model property' Umbraco property in the block's settings
         /// </summary>
         /// <param name="blockList">The block list to search</param>
@@ -64,7 +30,7 @@ namespace GovUk.Frontend.Umbraco.Models
             return RecursivelyFindBlock(blockList, x => x.Settings.GetProperty(PropertyAliases.ModelProperty)?.GetValue()?.ToString() == propertyName);
         }
 
-        private static BlockListItem? RecursivelyFindBlock(ReadOnlyCollection<BlockListItem> blockList, Func<BlockListItem, bool> matcher)
+        internal static BlockListItem? RecursivelyFindBlock(this ReadOnlyCollection<BlockListItem> blockList, Func<BlockListItem, bool> matcher)
         {
             if (blockList is null)
             {
